feature/ROOMS-14_login route names moved to router
